Extract generic drag and drop types
